MAINT: rename simdat.py to simdata.py
